+++ Project.toml
@@ -9,6 +9,8 @@ FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
 Gridap = "56d4f2e9-7ea1-5844-9cf6-b9c51ca7ce8e"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+MeshViz = "9ecf9c4f-6e5a-4b5e-83ae-06f2c7a661d8"
+Meshes = "eacbb407-ea5a-433e-ab97-5258b1ca43fa"
 
 [compat]
 Gridap = "0.16"
